fix(storybook): keep manifest docs helpers importable without the manifest
`.storybook/manifest-docs.ts` imported the gitignored
`custom-elements.json` at module level, so `script/manifest-docs.test.ts`
could not load in the `test-browser` CI job, which never runs
`npm run analyze`.

Split the pure logic into `.storybook/manifest-docs-core.ts`; the
manifest-bound bindings stay in `manifest-docs.ts`, which re-exports the
helper and the types so existing importers are unchanged.
